+++ griptape/structures/agent.py
@@ -47,10 +47,6 @@ class Agent(Structure):
         return super().add_tasks(*tasks)
 
     def try_run(self, *args) -> Agent:
-        self._execution_args = args
-
-        self.task.reset()
-
         self.task.execute()
 
         if self.conversation_memory and self.output is not None: